% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_indiv_df.R
\name{merge_indiv_df}
\alias{merge_indiv_df}
\title{Merge one data frame with many}
\usage{
merge_indiv_df(primary_df, df_names, column_name)
}
\arguments{
\item{primary_df}{data frame with info to merge with multiple other data frames.}

\item{df_names}{list of names of data frames to be merged from environment.}

\item{column_name}{column to use for merging.}
}
\value{
Multiple merged data frames.
}
\description{
This function uses a primary data frame and merges with secondary data frames.
}
\examples{
df1 <- data.frame(ID = c(1:10), num = "A")
df2 <- data.frame(ID = c(11:20), num = "B")
df_names <- ls()
primary_df <- data.frame(ID = c(1:20), num2 = c(100:119))
column_name <- "ID"
merge_indiv_df(primary_df, df_names, column_name)
df_list.test = mget(df_names)
# write_xlsx(df_list.test , "output.xlsx")
}
