% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_na.R
\name{remove_all_na_columns}
\alias{remove_all_na_columns}
\title{remove_all_na_columns}
\usage{
remove_all_na_columns(dataframe)
}
\value{
a data frame
}
\description{
This function removes columns with all NA values in a data frame
}
\examples{
df <- data.frame(
    ID = 1:5, Name = c("Alice", "Bob", "Charlie", "David", "Eve"),
    Score = c(85, 92, 78, 90, 88), AllNA1 = c(NA, NA, NA, NA, NA),
    AllNA2 = c(NA, NA, NA, NA, NA), Mixed = c(NA, 95, NA, 85, 80)
   )
remove_all_na_columns(df)
}
