% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceVarName.R
\name{replace_variable_names}
\alias{replace_variable_names}
\title{replace_variable_names}
\usage{
replace_variable_names(name_mapping)
}
\value{
a set of data frames
}
\description{
This function can be used to open all sheets from a list of excel files into separate data frames
}
\examples{
df.name <- c("a", "b", "c", "d", "e", "f", "g", "h")
df.name.new <- c("AmAcidPep", "BileAcids", "FattyAcids", "MicrobiomeGut",
                 "MSMLS_HIL", "MSMS_Plate", "MSMLS_RP", "OrganicAcids")
names(df.name.new) <- df.name
replace_variable_names(df.name.new)
}
