% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSheets.R
\name{save_all_sheets}
\alias{save_all_sheets}
\title{save_all_sheets}
\usage{
save_all_sheets(file)
}
\value{
a set of data frames
}
\description{
This function can be used to open all sheets from a list of excel files into separate data frames
}
\examples{
folder_path <- "myfolder/"
excel_files <- list.files(path = folder_path, pattern = "\\\\.xlsx$", full.names = TRUE)
walk(excel_files, save_all_sheets)
}
